requirements.txt
streamlit>=1.32.0
openai>=1.12.0
fpdf2>=2